rename robot.rs to robot_trait.rs to fix clippy warning
