+++ src/robot.rs
@@ -14,11 +14,11 @@ pub mod low_pass_filter;
 mod motion_generator_traits;
 pub mod panda;
 mod rate_limiting;
-pub mod robot;
 mod robot_control;
 pub mod robot_data;
 mod robot_impl;
 pub mod robot_state;
+pub mod robot_trait;
 pub mod robot_wrapper;
 pub(crate) mod service_types;
 pub(crate) mod types;
