+++ src/robot/robot_trait.rs
