+++ src/robot/robot_wrapper.rs
@@ -1,9 +1,9 @@
 use crate::exception::create_command_exception;
-use crate::robot::robot::PrivateRobot;
-use crate::robot::robot::Robot;
 use crate::robot::robot_control::RobotControl;
 use crate::robot::robot_data::RobotData;
 use crate::robot::robot_impl::RobotImplementation;
+use crate::robot::robot_trait::PrivateRobot;
+use crate::robot::robot_trait::Robot;
 use crate::robot::service_types::{
     SetCartesianImpedanceRequest, SetCollisionBehaviorRequest, SetEeToKRequest,
     SetGuidingModeRequest, SetJointImpedanceRequest, SetLoadRequest, SetNeToEeRequest,
